requirements.txt
streamlit
PyMuPDF
pdfplumber
python-docx
spacy
transformers
torch
sentencepiece
pandas
reportlab
Pillow
ollama
scipy
numpy
pytest
